refactor: delete `Preprocessor`
