+++ packages/shader-lab/src/index.ts
@@ -1,8 +1,7 @@
 export { ShaderLab } from "./ShaderLab";
 
 // #if _VERBOSE
-export { Preprocessor } from "./preprocessor";
-export { PpParser } from "./preprocessor/PpParser";
+export { PpParser } from "./preprocessor";
 export * from "./GSError";
 // #endif
