+++ packages/shader-lab/src/preprocessor/index.ts
@@ -1,1 +1,1 @@
-export * from "./Preprocessor";
+export * from "./PpParser";
