+++ tests/src/shader-lab/Preprocessor.test.ts
@@ -1,9 +1,9 @@
 import { testCaseList } from "./test-case";
 // @ts-ignore
 import { ShaderLib } from "@galacean/engine-core";
-import { Preprocessor, PpParser } from "@galacean/engine-shaderlab/verbose";
-import { describe, expect, it } from "vitest";
+import { PpParser } from "@galacean/engine-shaderlab/verbose";
 import { server } from "@vitest/browser/context";
+import { describe, expect, it } from "vitest";
 const { readFile } = server.commands;
 
 const includedSource = await readFile("test-case/included.txt");
@@ -12,7 +12,7 @@ ShaderLib["TEST"] = includedSource;
 describe("Preprocessor", () => {
   for (const testCase of testCaseList) {
     it(testCase.name, () => {
-      const out = Preprocessor.parse(testCase.source, [], []);
+      const out = PpParser.parse(testCase.source, [], []);
       expect(out).to.equal(testCase.compare);
     });
   }
